Move slot-wrapper tests next to the other project-settings tests
The slot wrappers belong to the settings page, so the tests live where
other settings tests do: `test/lightning_web/live/project_live_test.exs`
under a new `view-extension slot wrappers` describe.

The echo stub LiveComponent moves to `test/support/slot_echo_component.ex`,
following Lightning's existing pattern (`stub_usage_limiter.ex`,
`stub_rate_limiter.ex`). Drops the dedicated settings_slots_test.exs file
that didn't justify its own scaffolding.
